Add files via upload
I have a zip file which consists of input data as a csv file along with an image folder which consists of images in tiff format and to be used as an input. We have two modes of input i.e. text and images. To tackle this I have created image embeddings based on image ids with the help of Resnet50 model, used PCA for dimensionality reduction (as the image size is 256 x 256) so that we can use the image information as a feature in our model. I have combined the image embeddings and the training data. Next I have used XGboost to predict the target variables Temperature and Tint.

Please find the training and validation data from my drive link https://drive.google.com/drive/folders/1NVwcsnsF4AH02w1SV67Cbg8K4dpXJSTv?usp=drive_link